# don't unescape text in clear cases