After:

```
benchmarking decode/en/aeson/lazy
time                 1.630 ms   (1.606 ms .. 1.655 ms)
                     0.998 R²   (0.996 R² .. 0.999 R²)
mean                 1.638 ms   (1.626 ms .. 1.664 ms)
std dev              59.77 μs   (35.45 μs .. 104.0 μs)
variance introduced by outliers: 23% (moderately inflated)

benchmarking decode/jp/aeson
time                 2.382 ms   (2.359 ms .. 2.406 ms)
                     0.999 R²   (0.998 R² .. 0.999 R²)
mean                 2.404 ms   (2.392 ms .. 2.431 ms)
std dev              60.54 μs   (36.76 μs .. 112.9 μs)
variance introduced by outliers: 12% (moderately inflated)
```

before

```
benchmarking decode/en/aeson/lazy
time                 2.053 ms   (2.027 ms .. 2.077 ms)
                     0.998 R²   (0.996 R² .. 0.999 R²)
mean                 2.069 ms   (2.056 ms .. 2.098 ms)
std dev              62.97 μs   (34.19 μs .. 105.8 μs)
variance introduced by outliers: 17% (moderately inflated)

benchmarking decode/jp/aeson
time                 2.666 ms   (2.636 ms .. 2.699 ms)
                     0.998 R²   (0.997 R² .. 0.999 R²)
mean                 2.695 ms   (2.672 ms .. 2.735 ms)
std dev              103.1 μs   (64.72 μs .. 159.6 μs)
variance introduced by outliers: 22% (moderately inflated)
```

Even for japanese there seems to be small speedup, because record keys are ASCII