@phadej 2017-07-13
And with *typed decode* benchmark:

![screenshot from 2017-07-13 18-11-03](https://user-images.githubusercontent.com/51087/28173313-c26ee506-67f6-11e7-9ff4-4f66a4a06aa0.png)

decodeManual/direct/jp100 baseline & benchmarking decodeTH/direct/twitter100 severely inflated (I have to figure what causes so much noise on my laptop)