@bergmark 2017-07-18
Sorry for the delay! This looks great :-)